+++ packages/bots/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.9.4 - 2026-07-29
+
 ### Added
 
 - **Additive `platform_middleware` / `platform_conversation_middleware`